--- src/main/java/com/example/jungleroyal/domain/dto/UserDto.java
@@ -1,17 +1,0 @@
-package com.example.jungleroyal.domain.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class UserDto {
-    private Long id;
-    private String kakaoId;
-    private String username;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private LocalDateTime lastLoginAt;
-}
